+++ CHANGES.md
@@ -1,4 +1,5 @@
 # CHANGES
 
+- Feature: Ported LoopMediaPlayer class
 - Feature: Ported react-native-sound
 - Feature: Init